Student Login Verification
Checks if user has student role in MySQL and denies access if they do not. Also fixes a minor issue with a render which should've been a redirect. **You need to set up the users for this and the other two user types on MySQL to verify student role works, but root login allows access for the sake of testing.